// src/generated/prisma/models/Patient.ts
export type PatientWhereInput = {
  AND?: Prisma.PatientWhereInput | Prisma.PatientWhereInput[]
  OR?: Prisma.PatientWhereInput[]
  NOT?: Prisma.PatientWhereInput | Prisma.PatientWhereInput[]
  patientId?: Prisma.UuidFilter<"Patient"> | string
  phone?: Prisma.StringFilter<"Patient"> | string
  appointments?: Prisma.AppointmentListRelationFilter
  user?: Prisma.XOR<Prisma.UserScalarRelationFilter, Prisma.UserWhereInput>
}